chore: change the structure of base package
