+++ packages/base/.gitignore
@@ -1,0 +1,3 @@
+**/*.js
+**/*.d.ts
+**/*.js.map
