+++ packages/base/core/deducer.ts
