+++ packages/base/errors/credential-error.ts
