+++ packages/base/errors/index.ts
