+++ packages/base/package.json
@@ -3,17 +3,20 @@
   "version": "0.2.8",
   "description": "",
   "author": "plutolang",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
+  "main": "index.js",
+  "types": "index.d.ts",
   "files": [
-    "dist/"
+    "**/*.js",
+    "**/*.d.ts",
+    "**/*.js.map"
   ],
   "scripts": {
     "build": "tsc",
     "watch": "npm run build -- --watch",
     "lint": "eslint .",
     "test": "NODE_ENV=test vitest run --coverage --reporter=verbose",
-    "test:watch": "NODE_ENV=test vitest --coverage --reporter=verbose"
+    "test:watch": "NODE_ENV=test vitest --coverage --reporter=verbose",
+    "clean": "tsc --build --clean"
   },
   "devDependencies": {
     "@types/fs-extra": "^11.0.4",
