+++ packages/base/registry.ts
