+++ packages/base/simulator/index.ts
@@ -1,0 +1,2 @@
+export * from "./interfaces";
+export * from "./make-client";
