+++ packages/base/simulator/make-client.ts
@@ -1,28 +1,6 @@
+import { ServerRequest, ServerResponse, SimulatorCleint } from "./interfaces";
+
 const SIM_HANDLE_PATH = "/call";
-
-export interface IResourceInstance {
-  addEventHandler(op: string, args: string, fnResourceId: string): void;
-  setup(context: IContext): Promise<void>;
-  cleanup(): Promise<void>;
-}
-
-export interface IContext {
-  get serverUrl(): string;
-  findInstance(resourceName: string): IResourceInstance;
-}
-
-export interface ServerRequest {
-  resourceName: string;
-  op: string;
-  args: any[];
-}
-
-export interface ServerResponse {
-  readonly result?: any;
-  readonly error?: any;
-}
-
-export type SimulatorCleint = any;
 
 export function makeSimulatorClient(url: string, resourceName: string): SimulatorCleint {
   const get = (_target: any, op: string) => {
