+++ packages/base/utils/index.ts
@@ -2,3 +2,5 @@ export * from "./resource-id";
 export * from "./computing-unit";
 export * from "./configuration";
 export * from "./capturing-value";
+export * from "./provision";
+export * from "./platform";
