+++ packages/base/utils/resource-id.ts
