+++ packages/pluto-infra/src/utils/index.ts
@@ -1,1 +1,2 @@
 export { ImplClassMap } from "./impl-class-map";
+export { Sandbox } from "./sandbox";
