+++ packages/pluto-infra/test/sandbox.test.ts
@@ -1,7 +1,7 @@
-import fs from "fs";
-import path from "path";
+import * as fs from "fs";
+import * as path from "path";
 import { test, expect } from "vitest";
-import { Sandbox } from "./sandbox";
+import { Sandbox } from "../src/utils/sandbox";
 
 const code = `
 module.exports = {
